리뷰 커밋

+++ KOSTA_3_final/src/main/java/com/kosta/KOSTA_3_final/model/board/Review.java
@@ -35,7 +35,7 @@ public class Review {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name = "review_id")
-	int reviewId;
+	Long reviewId;
 	
 	@ManyToOne
 	@JoinColumn(name = "package_id")
